Migrate Binomial coefficient
- Move implementation to header file
- Add unit tests as cpp file
- Update readme
- Update build script
- Remove old files
